src/main/java/com/malicia/mrg/app/Ressources.java
package com.malicia.mrg.app;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.util.ArrayList;
import java.util.List;

public class Ressources {

    public static List<String> getResourceFiles(String path) throws IOException {
        List<String> filenames = new ArrayList<>();

        try (
                InputStream in = getResourceAsStream(path);
                BufferedReader br = new BufferedReader(new InputStreamReader(in))) {
            String resource;

            while ((resource = br.readLine()) != null) {
                filenames.add(resource);
            }
        }

        return filenames;
    }

    private static InputStream getResourceAsStream(String resource) {
        final InputStream in
                = getContextClassLoader().getResourceAsStream(resource);

        return in == null ? Ressources.class.getClassLoader().getResourceAsStream(resource) : in;
    }

    private static ClassLoader getContextClassLoader() {
        return Thread.currentThread().getContextClassLoader();
    }

    public static String listetostring(List<?> ele) {
        StringBuilder listeletostring = new StringBuilder();
        for (int i = 0; i < ele.size(); i++) {
            listeletostring.append("-->" + ele.get(i) + "\n");
        }
        return listeletostring.toString();
    }
}
